Add readBtn & fix removeBtn
Since the removeBtn was created in the global scope, only one existed on
the entire page. Now a new removeBtn is created within the createCard
function, and the eventListener has be move there as well. The actual
removeBook function remains unchanged, save for adding an important
e.target.id line.